Move OsDetector as a companion object of OsDetectorModule. Remove osDetectedMvnProperties from golden tests
